pyproject.toml
[project]
name = "elrif-auth-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Viacheslav Beatum",email = "podarkiprod@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.14,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pytest-mock = "^3.14.1"
aiosqlite = "^0.21.0"

